backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
whitenoise>=6.6
